Added variabales and calculations for fuel usage. Still need to implement fuel cost and multiplying to get yearly consumption

+++ Generators.h
@@ -4,9 +4,9 @@ ref class Generators
 	public:
 		double size;
 		int kwhYearly;
-		int fuelPerHour;
+		float fuelPerHour;
 		double cost;
-		Generators(int kwhYearly, int fuelPerHour, double cost, double size) {
+		Generators(int kwhYearly, float fuelPerHour, double cost, double size) {
 			this->size = size;
 			this->kwhYearly = kwhYearly;
 			this->fuelPerHour = fuelPerHour;
